Fix: DAO insert returns Long, JsItem type with default

+++ app/src/main/java/com/example/lotteryapp/data/MasterHistoryDao.kt
@@ -16,7 +16,7 @@ interface MasterHistoryDao {
     suspend fun getById(id: Int): MasterHistory?
     
     @Insert
-    suspend fun insert(master: MasterHistory): Int
+    suspend fun insert(master: MasterHistory): Long
     
     @Update
     suspend fun update(master: MasterHistory)
